requirements.txt
pyyaml
paho-mqtt
influxdb
